Completed mx autocompleter plugin.
Think *think* it works correctly for OTU.

THe plugin will return to the caller, anything in response_values hash.
So if you want otu[id] returned, then add
response_values=>{'otu[id]'=>1}

etc.

And a few other things...

On the autocompleter, if you put label_html - that will be used to
render the autocomplete items.
BUT -
label will be used as the thing which is in the text box.

So - you can have a fancy-schmancy autocomplete menu item piece of
content.
But when you select it, you can have a simple string to identify it.

Still a work in progress...  I think OTU is done. will send an email
later.

Thanks