Added jquery notifications. Passing flash in XHR headers back and forth. Those flash messages are pushed to the notification...
Also added some helpers so flash messages can be done like so:

notice "....."
warn   "....."
error  "....."

Have the tag page coming up, you try to create an empty tag, and it
shows an error and shakes.

I *don't* get what is going on if you successfully create a tag.
My plan though is to put all that logic into the link helper.
So when you create a link to make a tag on something, it will have to
have some info, like the class of item to update, or whatnot.
Will be intersting...