First cut. Think the tag create modal comes up and works correctly.
It will highlight the thing it tagged - passing in the html selector to
update.

I saw some code about having to update parts of the page with the new
tag... We'll figure that out when we get there... but I am thinking we'd
just implement another parameter that gets passed back - like
'html_update_selector' or something... and that will be updated with the
'tag_list'.

Someting like that.  new_tag_tag is now inserted everywhere.

Sorta coming along...