Completed deleting a tag from the OTU view
Not sure what else to do with this.

THe way I have always used RJS stuff is to add a known unique class to
the element when rendering, and then when it's destroyed -- we just fade
everything with that class.

So I have no idea if any of this works for other elements on the site...