Added mx-spinner logic.
This will allow one to put a 'lock' on specific parts of the page.
Not just the whole page at once.

It is also separate so that a modal can have an ajax spinner over it's
content...   Which is good for pages that use things like observe_field.

modal still uses it's own code since they are somewhat separate... the
modal one is fixed width, and a singleton.
This one is absolute, computed width, and can be multiples...