Completed adding password reset links and emails.
The password reset workflow emails out a reset password email.
It has a link to a page where you can set a new password for your
account.

Added tests for the token model and other such things.

Also - unearthed the preferences view (pulled out a bunch of broken
HTML?)  and plopped in two links to change email and change password -
since I could not find a link for them anywhere else on the site.

The also fixed the routes for those actions so you could hit them.