WIP on the inline form.
Was trying to do it inline with the HTML stored in the page, but that
just doesn't fly. It is too brittle and fragile for me. Trying to make
it an AJAX request and then see how that goes.  If it's too slow we'll
re-look at it inline an stored in the DOM.